requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
boto3==1.34.11
botocore==1.34.11
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
curl-cffi==0.5.10
dataclasses-json==0.6.3
distro==1.8.0
dnspython==2.4.2
duckduckgo_search==4.1.1
ecdsa==0.18.0
email-validator==2.1.0.post1
faiss-cpu==1.7.4
fastapi==0.105.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.354
langchain-community==0.0.8
langchain-core==0.1.6
langsmith==0.0.77
lxml==5.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.3.9
orjson==3.9.10
packaging==23.2
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rank-bm25==0.2.2
regex==2023.12.25
requests==2.31.0
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.24
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
